- Refractor User and Customer API - Created product API - Created payment instrument API - Started integrating Purchase flow

+++ jft-market/src/main/java/com/jft/market/api/ProductBean.java
@@ -9,13 +9,13 @@ import lombok.Setter;
 @Setter
 public class ProductBean {
 
-	private Integer id;
+	private Long id;
 	private String name;
 	private Integer price;
 	private String description;
 	private String features;
 
-	public ProductBean(Integer id, String name, Integer price, String description, String features) {
+	public ProductBean(Long id, String name, Integer price, String description, String features) {
 		this.id = id;
 		this.name = name;
 		this.price = price;
